Microsservico Cotacao/requirements.txt
# Framework web de alta performance
fastapi==0.111.0

# Servidor ASGI para rodar o FastAPI
uvicorn==0.30.1

# Ferramenta para validação de dados, usada pelo FastAPI
pydantic==2.7.4

# Cliente HTTP assíncrono para fazer requisições a APIs externas
httpx==0.27.0

# Biblioteca para gerenciar variáveis de ambiente do arquivo .env
python-dotenv==1.0.1

# Cliente para se conectar e interagir com o Redis
redis==5.0.1

# Suporte para GraphQL, caso você use essa interface
strawberry-graphql[fastapi]==0.231.0